Display any errors returned by `npm ls --production --parseable`
There are a variety of scenarios in which `npm ls` will return
a non-zero exit code and the output will include error messages.
In order to aid with debugging it is useful to be able to see
these error messages. The changes introduced in PR #105 had the
effect of silencing errors completely.

This change ensures that if the `npm ls` command was successful,
but the output contains some lines that aren't filepaths e.g.
loglevel info output, they will be stripped out, but that in the
case of `npm ls` returning an error, you will be able to see it
in the output displayed by `shellpromise`.